Changed colloquium announcement to a slideshow piece.
I decided that using the slideshow space for the colloquium announcement
made a lot more sense than the insertion into the text that I first used.
I've created a gimp template for colloquia.

I also commented out all the other slides so that the colloquium
announcement would always show. This can (and should) be changed
after the colloquium has passed.