Added a News Box
Added a News box on the right side of the intro text, to more or less mirror the slideshow.  This allows a visitor to easily see what the latest developments are, without waiting for the slideshow to scroll.  It's also easier to maintain, and more accessible.  I commented out the 2.4.8 image in the slideshow.