Add ref to use of pdl2 alongside perldl
pdl2 falls back to perldl so the documentation needs to
include it along with any references to the original
perldl so users know about it.  Once the full back-compatibility
is finished for pdl2, then the plan is to deprecate the perldl
usage.  Once the issues with having a pdl.exe and a pdl[.pl]
script that can interoperate seamlessly, then we can go to a
simpler pdl for all purposes.