Simplify the "Develop" page, and update sidebar links.
In Develop, I removed all the bits about Mercurial and Bazaar.  If people want to use alternates to Git, they can figure that out.  I also removed the instructions for how to use the various SCMs (we're not in the business of teaching Git), and instead included a link to a GitHub Git cheatsheet.

The sidebar links were updated to point to GitHub and not SourceForge.